clean up go files structure
